Birds of Azores archipelago

A dataset containing the age of the archipelago, number of species on mainland not present on the island, and each colonising clade with colonisation and branching times, endemicity status, missing species, and the type of species. Obtained and parsed from the relaxedDAISIE project strictly for illustration and test purposes

Usage

Azores

Format

An object of class list of length 18:

island_age

Age of the island

not_present

Number of mainland species that did not colonise the island

colonist_name

The name of the colonising clade

branching_times

The age of the island and the colonisation and subsequent branching times for a clade

stac

Endemicity status for a clade

missing_species

Number of species missing from a phylogeny of a clade

type1or2

Defines what type of species are in a clade, used for 2-type analysis

...

Source

Valente et al. (2020) doi: https://doi.org/10.1038/s41586-020-2022-5 https://github.com/joshwlambert/relaxedDAISIE