Runs a DAISIE simulation

Runs a DAISIE simulation

Usage

run_sim(daisie_data, model, lik_res, cond)

Arguments

daisie_data

A list, conforming to the DAISIE object format. Usually preprocessed by DAISIE::DAISIE_dataprep(), see it's documentation for details. Otherwise it may be generated via simulations, using DAISIE::DAISIE_sim_cr() and friends.

model

A string with model that should run. Models are as follows:

"cr_dd" Clade specific model - diversity dependent. All parameters free.

"cr_di" Clade specific model - diversity independent (K = Inf). All other parameters free.

"cr_dd_0laa" Clade specific model - diversity dependent and no anagenesis (laa fixed to zero). All other parameters free.

"cr_di_0laa" Clade specific model - diversity independent (K = Inf) and no anagenesis (laa fixed to zero). All other parameters free.

"cr_di_0lac" Clade specific model - diversity indendent (K = Inf) and no cladogenesis (lac fixed to zero). All other parameters free.

"cr_dd_0lac" Clade specific model - diversity dependent and no cladogenesis (lac fixed to zero). All other parameters free.

"rr_lac_di" Clade specific model - diversity independent (K = Inf) and relaxed cladogenesis. All other parameters free.

"rr_lac_dd" Clade specific model - diversity dependent and relaxed cladogenesis. All other parameters free.

"rr_mu_di" Clade specific model - diversity independent (K = Inf) and relaxed extinction. All other parameters free.

"rr_mu_dd" Clade specific model - diversity dependent and relaxed extinction. All other parameters free.

"rr_k" Clade specific model - diversity dependent and relaxed carrying capacity. All other parameters free.

"rr_gam_di" Clade specific model - diversity independent (K = Inf) and relaxed colonisation. All other parameters free.

"rr_gam_dd" Clade specific model - diversity dependent and relaxed colonisation. All other parameters free.

"rr_laa_di" Clade specific model - diversity independent (K = Inf) and relaxed anagenesis. All other parameters free.

"rr_laa_dd" Clade specific model - diversity dependent and relaxed anagenesis. All other parameters free.

"rr_mu_di_0lac" Clade specific model - diversity independent, relaxed extinction, and no cladogenesis (lac fixed to zero). All other parameters free.

"rr_mu_dd_0lac" Clade specific model - diversity dependent, relaxed extinction, and no cladogenesis (lac fixed to zero). All other parameters free.

"rr_k_0lac" Clade specific model - diversity dependent, relaxed carrying capacity, and no cladogenesis (lac fixed to zero). All other parameters free.

"rr_gam_di_0lac" Clade specific model - diversity independent, relaxed colonisation, and no cladogenesis (lac fixed to zero). All other parameters free.

"r_gam_dd_0lac" Clade specific model - diversity dependent, relaxed colonisation, and no cladogenesis (lac fixed to zero). All other parameters free.

"rr_laa_di_0lac" Clade specific model - diversity independent, relaxed anagenesis, and no cladogenesis (lac fixed to zero). All other parameters free.

"r_laa_dd_0lac" Clade specific model - diversity dependent, relaxed anagenesis, and no cladogenesis (lac fixed to zero). All other parameters free.

"rr_lac_di_0laa" Clade specific model - diversity independent, relaxed cladogenesis, and no anagenesis (laa fixed to zero)

"rr_lac_dd_0laa" Clade specific model - diversity dependent

"rr_mu_di_0laa" Clade specific model - diversity independent

"rr_mu_dd_0laa" Clade specific model - diversity dependent

"rr_k_0laa" Clade specific model - diversity dependent

"rr_gam_di_0laa" Clade specific model - diversity independent, relaxed colonisation, and no anagenesis (laa fixed to zero). All other parameters free

"rr_gam_dd_0laa" Clade specific model - diversity dependent, relaxed colonisation, and no anagenesis (laa fixed to zero). All other parameters free

lik_res

A data frame with results from a DAISIE maximum likelihood model.

cond

An integer specifying conditioning, as described in DAISIE::DAISIE_ML_CS().

Value

List with output from DAISIE simulation