Fills pars2 vector for most common DAISIE ML CS runs

Returns the pars2 vector needed for the most simple DAISIE_ML_CS() runs, to facilitate computing the loglikelihood in these cases. This function should only be assumed correct for the LL setup of the CS model, 1 type, constant rate no equilibrium models. Tolerances and other technical parameters need not be specified, as the returned values match the default values of DAISIE::DAISIE_ML(). The only arguments without default values are ddmodel and cond, as these are most often varied when checking model output. Regardless, all other default parameter values can be forced to meet specific needs of more complex models.

Usage

setup_std_pars2(
  res = 100,
  ddmodel = 11,
  cond = 0,
  verbose = 0,
  island_ontogeny = NA,
  eqmodel = 0,
  tol = c(1e-04, 1e-05, 1e-07),
  maxiter = 1000 * round((1.25)^5),
  x_E = 0.95,
  x_I = 0.98
)

Arguments

res

A numeric determining the resolution of the likelihood calculations, it sets the limit for the maximum number of species for which a probability must be computed, which must be larger than the size of the largest clade.

ddmodel

Sets the model of diversity-dependence:

ddmodel = 0: no diversity dependence

ddmodel = 1: linear dependence in speciation rate

ddmodel = 11: linear dependence in speciation rate and in immigration rate

ddmodel = 2: exponential dependence in speciation rate

ddmodel = 21: exponential dependence in speciation rate and in immigration rate

cond

An integer specifying conditioning, as described in DAISIE::DAISIE_ML_CS().

verbose

In simulation and dataprep functions a logical, Default = TRUE gives intermediate output should be printed. For ML functions a numeric determining if intermediate output should be printed, Default = 0 does not print, verbose = 1 prints intermediate output of the parameters and loglikelihood, verbose = 2 means also intermediate progress during loglikelihood computation is shown.

island_ontogeny

In DAISIE_sim_time_dep(), DAISIE_ML_CS and plotting a string describing the type of island ontogeny. Can be "const", "beta" for a beta function describing area through time.
In all other functions a numeric describing the type of island ontogeny. Can be 0 for constant, 1 for a beta function describing area through time. In ML functions island_ontogeny = NA assumes constant ontogeny. Time dependent estimation is not yet available as development is still ongoing. Will return an error if called in that case.

eqmodel

Sets the equilibrium constraint that can be used during the likelihood optimization. Only available for datatype = 'single'.

eqmodel = 0 : no equilibrium is assumed
eqmodel = 13 : near-equilibrium is assumed on endemics using deterministic equation for endemics and immigrants. Endemics must be within x_E of the equilibrium value
eqmodel = 15 : near-equilibrium is assumed on endemics and immigrants using deterministic equation for endemics and immigrants. Endemics must be within x_E of the equilibrium value, while non-endemics must be within x_I of the equilibrium value.

tol

Sets the tolerances in the optimization. Consists of:

reltolx - relative tolerance of parameter values in optimization.

reltolf - relative tolerance of function value in optimization.

abstolx - absolute tolerance of parameter values in optimization.

maxiter

Sets the maximum number of iterations in the optimization.

x_E

Sets the fraction of the equlibrium endemic diversity above which the endemics are assumed to be in equilibrium; only active for eqmodel = 13 or 15.

x_I

Sets the fraction of the equlibrium non-endemic diversity above which the system is assumed to be in equilibrium; only active for eqmodel = 15.

Value

A numeric vector of length 12 containing pars2 for DAISIE::DAISIE_ML()

Author

Pedro Santos Neves

Examples

std_pars2 <- setup_std_pars2()